overlay2: factor checks out into overlayutils
